Extract image cacheing test cases into the new `CacheImageUseCaseTests` class
