+++ FeedModules/Core Layer/Tests/FeedCoreModuleTests/Feed Cache/CacheImageUseCaseTests.swift
@@ -1,0 +1,39 @@
+//
+//  CacheImageUseCaseTests.swift
+//  FeedCoreModule
+//
+//  Created by Omran Khoja on 3/10/22.
+//
+
+import XCTest
+import FeedCoreModule
+
+class CacheFeedImageDataUseCaseTests: XCTestCase {
+    
+    func test_init_doesNotMessageStoreUponCreation() {
+        let (_, store) = makeSUT()
+        
+        XCTAssertTrue(store.receivedMessages.isEmpty)
+    }
+    
+    func test_saveImageDataForURL_requestsImageDataInsertionForURL() {
+        let (sut, store) = makeSUT()
+        let url = anyURL()
+        let data = anyData()
+        
+        sut.save(data, for: url) { _ in }
+        
+        XCTAssertEqual(store.receivedMessages, [.insert(data: data, for: url)])
+    }
+    
+    // MARK: - Helpers
+    
+    private func makeSUT(file: StaticString = #file, line: UInt = #line) -> (sut: LocalImageLoader, store: ImageDataStoreSpy) {
+        let store = ImageDataStoreSpy()
+        let sut = LocalImageLoader(store: store)
+        trackForMemoryLeaks(store, file: file, line: line)
+        trackForMemoryLeaks(sut, file: file, line: line)
+        return (sut, store)
+    }
+    
+}
